refactor(sandbox): split persistence helpers
